windows: Move rawinput stuff into its own file.
