+++ Windows/KeyboardDevice.h
@@ -4,6 +4,8 @@
 #include "base/mutex.h"
 #include "InputDevice.h"
 
+extern std::map<int, int> windowsTransTable;
+
 class KeyboardDevice : public InputDevice {
 public:
 	virtual int UpdateState(InputState &input_state);
